+++ lib/action_view/attributes_and_token_lists/attribute_merger.rb
@@ -1,0 +1,36 @@
+module ActionView
+  module AttributesAndTokenLists
+    class AttributeMerger < ActiveSupport::OptionMerger
+      def initialize(view_context, context, options)
+        @view_context = view_context
+        super context, @view_context.tag.attributes(options.to_hash.with_indifferent_access)
+      end
+
+      def with_attributes(options, &block)
+        attribute_merger = AttributeMerger.new @view_context, @context, @options.merge(options)
+
+        if block.nil?
+          attribute_merger
+        else
+          block.arity.zero? ? attribute_merger.instance_eval(&block) : block.call(attribute_merger)
+        end
+      end
+
+      def tag(name = nil, options = nil, open = false, escape = true)
+        if [ name, options ].all?(&:nil?)
+          AttributeMerger.new(@view_context, @view_context.tag, @options)
+        else
+          @view_context.tag(name, @options.merge(options.to_h))
+        end
+      end
+
+      def to_hash
+        @options.to_hash
+      end
+
+      def to_h
+        @options.to_h
+      end
+    end
+  end
+end